complete firebase perf

+++ proj.android_studio/ee-x-firebase-performance/src/main/java/com/ee/firebase/performance/FirebasePerformanceTrace.java
@@ -77,7 +77,7 @@ public class FirebasePerformanceTrace {
                 assert dict != null;
 
                 String key = (String) dict.get(k__key);
-                Long value = (Long) dict.get(k__value);
+                Integer value = (Integer) dict.get(k__value);
                 incrementMetric(key, value);
                 return "";
             }
